moves preview table components into subdirectories
