+++ src/lib/components/preview-table/core/Cell.svelte
@@ -1,16 +1,16 @@
 <script>
+  import { FormattedDataType } from "$lib/components/data-types";
+  import notificationStore from "$lib/components/notifications";
+  import Shortcut from "$lib/components/tooltip/Shortcut.svelte";
+  import StackingWord from "$lib/components/tooltip/StackingWord.svelte";
+  import Tooltip from "$lib/components/tooltip/Tooltip.svelte";
+  import TooltipContent from "$lib/components/tooltip/TooltipContent.svelte";
+  import TooltipShortcutContainer from "$lib/components/tooltip/TooltipShortcutContainer.svelte";
+  import TooltipTitle from "$lib/components/tooltip/TooltipTitle.svelte";
   import { INTERVALS, TIMESTAMPS } from "$lib/duckdb-data-types";
   import { formatDataType } from "$lib/util/formatters";
   import { createShiftClickAction } from "$lib/util/shift-click-action";
   import { createEventDispatcher } from "svelte";
-  import { FormattedDataType } from "../data-types";
-  import notificationStore from "../notifications";
-  import Shortcut from "../tooltip/Shortcut.svelte";
-  import StackingWord from "../tooltip/StackingWord.svelte";
-  import Tooltip from "../tooltip/Tooltip.svelte";
-  import TooltipContent from "../tooltip/TooltipContent.svelte";
-  import TooltipShortcutContainer from "../tooltip/TooltipShortcutContainer.svelte";
-  import TooltipTitle from "../tooltip/TooltipTitle.svelte";
 
   export let row;
   export let column;
