+++ src/lib/components/preview-table/core/Row.svelte
