+++ src/lib/components/preview-table/types.d.ts
@@ -1,0 +1,1 @@
+export type HeaderPosition = "top" | "left" | "top-left";
